requirements.txt
anyio==4.6.2.post1
ariadne==0.23.0
blinker==1.9.0
click==8.1.7
exceptiongroup==1.2.2
flask==3.1.0
Flask-Cors==5.0.0
flask-sqlalchemy==3.1.1
graphql-core==3.2.5
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing-extensions==4.12.2
werkzeug==3.1.3
zipp==3.21.0
